Switched to double precision numbers as the basic sample type.

+++ src/global.h
@@ -49,7 +49,7 @@ namespace Aquila
     /**
      * Sample value type.
      */
-    typedef int SampleType;
+    typedef double SampleType;
 
     /**
      * Sample frequency type.
